Feature: Add `bf2stats` `2.2.0` variants
This is for those who want to run the classic `bf2stats` `2.2.0` as-is. Note that as of `bf2stats` `2.2.0`, the python files of the BF2 server does not support `Transfer-Encoding: chunked` responses from the ASP web server, so the ASP web server must respond without that HTTP header. See: https://github.com/startersclan/bf2stats/pull/37